requirements.txt
Django==2.1.7
djangorestframework==3.9.2
mysqlclient==1.4.2.post1
ipython==7.3.0
